Explanation of Revamp
This is a further explanation of what I did: I changed the Queue to QueueLL for linked list, and I restructured it to how Dr. Wolff has the linked list in her examples. The way we had it before was this confusing combination between a linked list and a queue. I created an update queue method in that class, which updates the queue based on the time and maoves the customers in and out of the queue accordingly. Instead of having to do a bunch of work focussing on which customer in the queue is being served, I changed the code around the idea to assume the first item was the customer in service. This way, you add customers to the end of the linked list for the queue and when the time is up on the first customer in the queue, you delete that customer from the queue, which makes the next customer the first in line. I changed the customer creator so that it doesn't use unecessary fields, and it also sorts and id's the customers within that class instead of another class. I scrapped the old simulator and fed it only the customers and the three queues as the fields. The new algorithm is that while all the customers are not served, it first, adds all the arriving customers to the shortest queue, updates each queue based on the time and adds the amount of customers served if a customer is done being served, and then adds to the currentTime by one and repeats.