Began Checkpoint D
Created a createFullQueue and createSelfQueue methods in the SimDriver class. These are passed in a number and new queues are created for the number passed in and appended to an AL which is returned. Not completely working as we wanted queues that were 26 or greater to be AA BB CC and so forth but instead its AA AB AC