changes everyehere
Made visualSimMenu work with simulating the data. (It is basically both SimulatorDriver and Simulator) It outputs the simulation on screen but goes to fast to see anything (not sure how to slow it down)
Added a SimulationSettingsClass which basically is used to transport the settings amongst the classes.
Added more functionality between visualSimMenu and  SimSetUpMenu via they SimSetUpMenu can jump straight to visualSimMenu when pressing StartSim. Many more changes BUT STILL NEED TO FIGURE OUT DATA TABLE PAGE (i.e outputting the results to it in the JPane)