Connection
I connected all the menu's together, changed all the menu's so that the image icon's work in the csc223Giraffes folder, changed the visual sim a tiny bit so the buttons are in a better place and so that it sends the Arraylist of customer data back to the menu manager. After running the sim setup, the data table will have the correct data now.